Added install paths to scripts and removed white spaces

+++ extras/scripts/hypre_setup.sh
@@ -7,6 +7,11 @@ BUILD_HYPRE=0
 ROCM_VERSION=6.0
 SUDO="sudo"
 DEB_FRONTEND="DEBIAN_FRONTEND=noninteractive"
+AMDGPU_GFXMODEL_INPUT=""
+USE_SPACK=1
+HYPRE_VERSION="2.32.0"
+HYPRE_PATH=/opt/rocmplus-${ROCM_VERSION}/hypre
+HYPRE_PATH_INPUT=""
 
 if [  -f /.singularity.d/Singularity ]; then
    SUDO=""
@@ -20,11 +25,14 @@ DISTRO_VERSION=`cat /etc/os-release | grep '^VERSION_ID' | sed -e 's/VERSION_ID=
 usage()
 {
    echo "Usage:"
-   echo "  --module-path [ MODULE_PATH ] default /etc/lmod/modules/misc/hypre"
-   echo "  --rocm-version [ ROCM_VERSION ] default $ROCM_VERSION"
-   echo "  --amdgpu-gfxmodel [ AMDGPU-GFXMODEL ] default autodetected"
-   echo "  --build-hypre [ BUILD_HYPRE ] default is 0"
-   echo "  --help: this usage information"
+   echo "  --module-path [ MODULE_PATH ] default is $MODULE_PATH "
+   echo "  --install-path [ HYPRE_PATH_INPUT ] default is $HYPRE_PATH "
+   echo "  --rocm-version [ ROCM_VERSION ] default $ROCM_VERSION "
+   echo "  --amdgpu-gfxmodel [ AMDGPU_GFXMODEL_INPUT ] default autodetected "
+   echo "  --hypre-version [ HYPRE_VERSION ] default is $HYPRE_VERSION "
+   echo "  --use-spack [ USE_SPACK ] default is $USE_SPACK "
+   echo "  --build-hypre [ BUILD_HYPRE ] default is 0 "
+   echo "  --help: this usage information "
    exit 1
 }
 
@@ -46,12 +54,17 @@ do
    case "${1}" in
       "--amdgpu-gfxmodel")
           shift
-          AMDGPU_GFXMODEL=${1}
+          AMDGPU_GFXMODEL_INPUT=${1}
           reset-last
           ;;
       "--build-hypre")
           shift
           BUILD_HYPRE=${1}
+          reset-last
+          ;;
+      "--hypre-version")
+          shift
+          HYPRE_VERSION=${1}
           reset-last
           ;;
       "--help")
@@ -60,6 +73,15 @@ do
       "--module-path")
           shift
           MODULE_PATH=${1}
+          reset-last
+          ;;
+      "--install-path")
+          shift
+          INSTALL_PATH_INPUT=${1}
+          reset-last
+      "--use-spack")
+          shift
+          USE_SPACK=${1}
           reset-last
           ;;
       "--rocm-version")
@@ -78,13 +100,30 @@ do
    shift
 done
 
+if [ "${HYPRE_PATH_INPUT}" != "" ]; then
+   HYPRE_PATH=${HYPRE_PATH_INPUT}
+else
+   # override path in case ROCM_VERSION has been supplied as input
+   HYPRE_PATH=/opt/rocmplus-${ROCM_VERSION}/hypre
+fi
+
 echo ""
 echo "==================================="
 echo "Starting HYPRE Install with"
 echo "ROCM_VERSION: $ROCM_VERSION"
 echo "BUILD_HYPRE: $BUILD_HYPRE"
+echo "HYPRE_VERSION: $HYPRE_VERSION"
+echo "HYPRE_PATH: $HYPRE_PATH"
+echo "MODULE_PATH: $MODULE_PATH"
 echo "==================================="
 echo ""
+
+if [[ "$AMDGPU_GFXMODEL_INPUT" != "" ]]; then
+   AMDGPU_GFXMODEL=$AMDGPU_GFXMODEL_INPUT
+else
+   AMDGPU_GFXMODEL=`rocminfo | grep gfx | sed -e 's/Name://' | head -1 |sed 's/ //g'`
+fi
+
 
 AMDGPU_GFXMODEL_STRING=`echo ${AMDGPU_GFXMODEL} | sed -e 's/;/_/g'`
 CACHE_FILES=/CacheFiles/${DISTRO}-${DISTRO_VERSION}-rocm-${ROCM_VERSION}-${AMDGPU_GFXMODEL_STRING}
@@ -124,7 +163,11 @@ else
 
       cd /tmp
 
-      export HYPRE_PATH=/opt/rocmplus-${ROCM_VERSION}/hypre
+      # don't use sudo if user has write access to install path
+      if [ -w ${HYPRE_PATH} ]; then
+         SUDO=""
+      fi
+
       ${SUDO} mkdir -p ${HYPRE_PATH}
 
       if [[ "${USER}" != "root" ]]; then
@@ -133,48 +176,61 @@ else
 
       # ------------ Installing HYPRE
 
-      git clone https://github.com/spack/spack.git
+      if [[ $USE_SPACK == 1 ]]; then
 
-      # load spack environment
-      source spack/share/spack/setup-env.sh
+         git clone https://github.com/spack/spack.git
 
-      # find already installed libs for spack
-      spack external find
+         # load spack environment
+         source spack/share/spack/setup-env.sh
 
-      # change spack install dir for Hypre
-      ${SUDO} sed -i 's|$spack/opt/spack|'"${HYPRE_PATH}"'|g' spack/etc/spack/defaults/config.yaml
+         # find already installed libs for spack
+         spack external find
 
-      # open permissions to use spack to install hypre
-      if [[ "${USER}" != "root" ]]; then
-         ${SUDO} chmod -R a+rwX ${HYPRE_PATH}
+         # change spack install dir for Hypre
+         ${SUDO} sed -i 's|$spack/opt/spack|'"${HYPRE_PATH}"'|g' spack/etc/spack/defaults/config.yaml
+
+         # open permissions to use spack to install hypre
+         if [[ "${USER}" != "root" ]]; then
+            ${SUDO} chmod -R a+rwX ${HYPRE_PATH}
+         fi
+
+         # install hypre with spack
+         #spack install hypre+rocm+rocblas+unified-memory
+         spack install hypre@$HYPRE_VERSION+rocm@$ROCM_VERSION+rocblas@$ROCM_VERSION+unified-memory+gpu-aware-mpi amdgpu_target=$AMDGPU_GFXMODEL
+
+         # get hypre install dir created by spack
+         HYPRE_PATH=`spack find -p hypre | awk '{print $2}' | grep opt`
+
+         ${SUDO} rm -rf spack
+
+      else
+
+         echo " HYPRE build only enabled with Spack at the moment "
+         echo " Set --use-spack 1 "
+         exit 1
+
       fi
 
-      # install hypre with spack
-      #spack install hypre+rocm+rocblas+unified-memory
-      spack install hypre+rocm+rocblas+unified-memory+gpu-aware-mpi amdgpu_target=gfx942
-
-
-      # get hypre install dir created by spack
-      HYPRE_PATH=`spack find -p hypre | awk '{print $2}' | grep opt`
-
-      ${SUDO} rm -rf spack
-
       if [[ "${USER}" != "root" ]]; then
-         ${SUDO} find ${HYPRE_PATH} -type f -execdir chown root:root "{}" +
+            ${SUDO} find ${HYPRE_PATH} -type f -execdir chown root:root "{}" +
       fi
       if [[ "${USER}" != "root" ]]; then
          ${SUDO} chmod go-w ${HYPRE_PATH}
       fi
 
       module unload rocm/${ROCM_VERSION}
+      module unload openmpi
 
    fi
 
    # Create a module file for hypre
+   if [ ! -w ${MODULE_PATH} ]; then
+      SUDO="sudo"
+   fi
    ${SUDO} mkdir -p ${MODULE_PATH}
 
    # The - option suppresses tabs
-   cat <<-EOF | ${SUDO} tee ${MODULE_PATH}/2.32.0.lua
+   cat <<-EOF | ${SUDO} tee ${MODULE_PATH}/$HYPRE_VERSION.lua
 	whatis("HYPRE - solver package")
 
 	local base = "${HYPRE_PATH}"
